Add new HTML file for MagPhase Chart and update delay handlers in main.js
A new HTML file has been added to present a 'MagPhase Chart', including the necessary scripts for visualization and interaction with it. In the main.js file, the handling of delay input has been updated for a more precise control. Changes also include modifications to multiple functions for optimisation and fix polarity inversion implementation.